src/creditunder/pyproject.toml
[project]
name = "creditunder"
version = "0.1.0"
description = "AI Credit Underwriting Platform - processor and workbench API"
requires-python = ">=3.14"
dependencies = [
    "aiokafka>=0.14.0",
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi>=0.136.1",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "openai>=2.33.0",
    "psycopg2-binary>=2.9.12",
    "pydantic>=2.13.3",
    "pydantic-ai[openai]>=1.89.0",
    "pydantic-settings>=2.14.0",
    "pyyaml>=6.0.2",
    "python-multipart>=0.0.27",
    "sqlalchemy[asyncio]>=2.0.49",
    "sse-starlette>=3.4.1",
    "structlog>=25.5.0",
    "uvicorn[standard]>=0.46.0",
    "weasyprint>=68.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/creditunder"]
